+++ R/ModelSEIR.R
@@ -7,8 +7,6 @@
 #' @param incubation_days Numeric scalar greater than 0. Average number of
 #' incubation days.
 #' @param recovery_rate Numeric scalar between 0 and 1. Rate of recovery_rate from virus.
-#' @param x Object of class SEIR.
-#' @param ... Currently ignore.
 #' @export
 #' @family Models
 #' @aliases epiworld_seir
@@ -45,19 +43,16 @@ ModelSEIR <- function(
     incubation_days,
     recovery_rate
     ) {
+  # Check input parameters
+  stopifnot_string(name)
+  stopifnot_double(prevalence)
+  stopifnot_double(transmission_rate)
+  stopifnot_double(incubation_days)
+  stopifnot_double(recovery_rate)
 
   structure(
     ModelSEIR_cpp(name, prevalence, transmission_rate, incubation_days, recovery_rate),
     class = c("epiworld_seir", "epiworld_model")
   )
 
-}
-
-#' @rdname ModelSEIR
-#' @param main Title of the plot
-#' @returns The `plot` function returns a plot of the SEIR model of class
-#' [epiworld_model].
-#' @export
-plot.epiworld_seir <- function(x, main = get_name(x), ...) { # col = NULL
-  plot_epi(x, main = main, ...)
 }