+++ R/ModelSEIRCONN.R
@@ -12,8 +12,6 @@
 #' @param incubation_days Numeric scalar greater than 0. Average number of
 #' incubation days.
 #' @param recovery_rate Numeric scalar between 0 and 1. Probability of recovery_rate.
-#' @param x Object of class SEIRCONN.
-#' @param ... Currently ignore.
 #' @param n Number of individuals in the population.
 #' @export
 #' @family Models
@@ -49,9 +47,22 @@
 #' plot(model_seirconn)
 #' @seealso epiworld-methods
 ModelSEIRCONN <- function(
-    name, n, prevalence, contact_rate, transmission_rate,
-    incubation_days, recovery_rate
+    name,
+    n,
+    prevalence,
+    contact_rate,
+    transmission_rate,
+    incubation_days,
+    recovery_rate
     ) {
+  # Check input parameters
+  stopifnot_string(name)
+  stopifnot_int(n)
+  stopifnot_double(prevalence)
+  stopifnot_double(contact_rate)
+  stopifnot_double(transmission_rate)
+  stopifnot_double(incubation_days)
+  stopifnot_double(recovery_rate)
 
   structure(
     ModelSEIRCONN_cpp(name, n, prevalence, contact_rate,
@@ -59,13 +70,4 @@ ModelSEIRCONN <- function(
     class = c("epiworld_seirconn", "epiworld_model")
   )
 
-}
-
-#' @rdname ModelSEIRCONN
-#' @export
-#' @returns The `plot` function returns a plot of the SEIRCONN model of class
-#' [epiworld_model].
-#' @param main Title of the plot.
-plot.epiworld_seirconn <- function(x, main = get_name(x), ...) { # col = NULL
-  plot_epi(x, main = main, ...)
 }