+++ R/ModelSIS.R
@@ -7,8 +7,6 @@
 #' @param transmission_rate Numeric scalar between 0 and 1. Virus's rate of
 #' infection.
 #' @param recovery_rate Numeric scalar between 0 and 1. Rate of recovery from virus.
-#' @param x Object of class SIS.
-#' @param ... Currently ignore.
 #' @export
 #' @family Models
 #' @returns
@@ -36,22 +34,19 @@
 #'
 #' @seealso epiworld-methods
 ModelSIS <- function(
-    name, prevalence, transmission_rate, recovery_rate) {
+    name,
+    prevalence,
+    transmission_rate,
+    recovery_rate) {
+  # Check input parameters
+  stopifnot_string(name)
+  stopifnot_double(prevalence)
+  stopifnot_double(transmission_rate)
+  stopifnot_double(recovery_rate)
 
   structure(
     ModelSIS_cpp(name, prevalence, transmission_rate, recovery_rate),
     class = c("epiworld_sis", "epiworld_model")
   )
 
-}
-
-
-#' @rdname ModelSIS
-#' @export
-#' @returns
-#' - The `plot` function returns a plot of the SIS model of class
-#' [epiworld_model].
-#' @param main Title of the plot.
-plot.epiworld_sis <- function(x, main = get_name(x), ...) { # col = NULL
-  plot_epi(x, main = main, ...)
 }