+++ inst/include/epiworld/epiworld.hpp
@@ -12,13 +12,16 @@
 #include <cstdint>
 #include <algorithm>
 #include <regex>
+#include <sstream>
+#include <iomanip>
+#include <set>
 
 #ifndef EPIWORLD_HPP
 #define EPIWORLD_HPP
 
 /* Versioning */
 #define EPIWORLD_VERSION_MAJOR 0
-#define EPIWORLD_VERSION_MINOR 6
+#define EPIWORLD_VERSION_MINOR 8
 #define EPIWORLD_VERSION_PATCH 1
 
 static const int epiworld_version_major = EPIWORLD_VERSION_MAJOR;
@@ -32,6 +35,9 @@ namespace epiworld {
 
     #include "misc.hpp"
     #include "progress.hpp"
+
+    #include "modeldiagram-bones.hpp"
+    #include "modeldiagram-meat.hpp"
 
     #include "math/distributions.hpp"