+++ inst/tinytest/test-sir.R
@@ -4,9 +4,9 @@
 test_tmat_matches_expected <- function(tmat) {
   tmat_expected <- structure(
     c(
-      0.961843, 0, 0,
-      0.03815696, 0.69985167, 0,
-      0, 0.3001483, 1
+      0.963299980893923, 0, 0,
+      0.0367000191060766, 0.700012995826896, 0, 
+      0, 0.299987004173104, 1
     ),
     dim = c(3L, 3L),
     dimnames = list(
@@ -59,3 +59,45 @@ runtime_noqueuing <- system.time(run(sir_0, ndays = 50, seed = 1912))
 queuing_on(sir_0)
 runtime_queuing <- system.time(run(sir_0, ndays = 50, seed = 1912))
 expect_true(runtime_queuing["elapsed"] < runtime_noqueuing["elapsed"])
+
+# Check functions fail with invalid inputs -------------------------------------
+good_name <- "A Virus"
+good_prevalence <- 0.01
+good_transmission_rate <- 0.9
+good_recovery_rate <- 0.3
+
+bad_name <- NA
+bad_prevalence <- NA
+bad_transmission_rate <- NA
+bad_recovery_rate <- NA
+
+expected_error_msg_str <- "must be a string"
+expected_error_msg_double <- "must be a double"
+
+expect_error(test_model <- ModelSIR(
+  name = bad_name,
+  prevalence = good_prevalence,
+  transmission_rate = good_transmission_rate,
+  recovery_rate = good_recovery_rate
+), expected_error_msg_str)
+
+expect_error(test_model <- ModelSIR(
+  name = good_name,
+  prevalence = bad_prevalence,
+  transmission_rate = good_transmission_rate,
+  recovery_rate = good_recovery_rate
+), expected_error_msg_double)
+
+expect_error(test_model <- ModelSIR(
+  name = good_name,
+  prevalence = good_prevalence,
+  transmission_rate = bad_transmission_rate,
+  recovery_rate = good_recovery_rate
+), expected_error_msg_double)
+
+expect_error(test_model <- ModelSIR(
+  name = good_name,
+  prevalence = good_prevalence,
+  transmission_rate = good_transmission_rate,
+  recovery_rate = bad_recovery_rate
+), expected_error_msg_double)